mmc: sunxi: Add support for Allwinner D1.
