+++ drivers/mmc/Kconfig
@@ -750,7 +750,7 @@ config ZYNQ_HISPD_BROKEN
 
 config MMC_SUNXI
 	bool "Allwinner sunxi SD/MMC Host Controller support"
-	depends on ARCH_SUNXI
+	depends on ARCH_SUNXI || TARGET_SUNXI
 	default y
 	help
 	  This selects support for the SD/MMC Host Controller on
